feat: hardware commissioning, display polish, and docs
Fixes and improvements made during initial hardware bring-up:

Connectivity
- Add USB CDC flags for ESP32-C3 native USB serial output
- Wait for WiFi before MQTT to prevent DNS failures at boot
- Rate-limit WiFi reconnect to avoid stack error spam
- Set DHCP hostname to 'gps-skyview'

Display
- Fix GC9A01A color inversion (panel config invert=true)
- Add sprite double-buffering (8-bit, 57KB) to eliminate flicker
- Add fillScreen before fillCircle to prevent ghost satellite buildup
- Fix el/az/ss JSON parsing (GPSD sends floats, not ints)
- Add Zulu time readout from TPV message (HH:MM:SS Z)
- Add minor elevation rings at 15/45/75 degrees (dimmer grey)
- Switch trails to light grey mono color
- Quadrant-aware PRN label placement (inward-facing to stay in bezel)
- Cardinal labels tuned to kR+4 offset

Tests
- Add float-payload test for sky parser (guards float regression)
- Update TPV tests for new two-arg signature and Z suffix

Docs
- Add README.md with hardware, wiring, setup, and architecture
- Add design spec and implementation plan to claude-working-documents/
- Add LCD reference documents and wiring diagram

+++ src/model/SkyMessage.cpp
@@ -1,4 +1,5 @@
 // Claude Generated: version 1 - ArduinoJson parser for gps_monitor/sky MQTT payload
+// Claude Generated: version 2 - Parse el/az/ss as float then cast; GPSD sends them as JSON floats
 #include "model/SkyMessage.h"
 #include <ArduinoJson.h>
 
@@ -10,11 +11,13 @@ bool parseSky(const char* json, SkyData& out) {
     out.satUsed    = doc["sat_used"]    | 0;
     out.satVisible = doc["sat_visible"] | 0;
     for (JsonObject s : doc["satellites"].as<JsonArray>()) {
+        // el/az/ss come from GPSD as JSON floats (e.g. 72.0); | 0.0f reads them correctly,
+        // then cast to int for our struct. Using | 0 (int) would fail the is<int>() check.
         out.satellites.push_back(Satellite{
             s["prn"]  | 0,
-            s["el"]   | 0,
-            s["az"]   | 0,
-            s["ss"]   | 0,
+            (int)(s["el"]   | 0.0f),
+            (int)(s["az"]   | 0.0f),
+            (int)(s["ss"]   | 0.0f),
             s["used"] | false
         });
     }
